Modify the passing of namespaces and attributes
The namespaces and attributes are now provided within a Hash instead of
within an Array. This allows the namespace prefixes found by libxml2 to
be added to the Hash.

The namespaces are provided within a Hash to unify the passing of
namespaces and attributes to the class who mixes in
LibXML::XML::SaxParser::Callbacks.

The values of a namespace are retrieved with the keys
 - :prefix (the prefix of the default namespace is the empty string)
 - :uri

The values of an attribute are retrieved with the keys
 - :prefix
 - :localname
 - :value